move lists stuff into its own file, where it can rot
